requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.7.0
distro==1.9.0
Django==5.1.3
fonttools==4.55.0
frozenlist==1.5.0
gigachat==0.1.36
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
kombu==5.4.2
multidict==6.1.0
openai==1.55.1
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydyf==0.11.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
yarl==1.18.0
zopfli==0.2.3.post1
